added SeasonStats module and test, moved #most_hits, #least_hits to module and module test
